clients/bot/requirements.txt
grpcio>=1.34.1
pika>=1.3.2
protobuf>=3.15.8
transitions>=0.9.0
grpcio-tools>=1.54.2
requests>=2.31.0
